+++ src/components/modals/Theme.tsx
@@ -1,22 +1,24 @@
 import { useEffect, useContext } from 'react'
+import { useTranslation } from 'react-i18next'
 import List from '@mui/material/List'
 import ListItemButton from '@mui/material/ListItemButton'
 import ListItemText from '@mui/material/ListItemText'
 
 import { AppBarTitleContext, ThemeContext } from '~/context'
-import { PaletteMode } from '~/styles/theme'
+import { ThemePreference } from '~/styles/theme'
 import Modal from '~/components/Modal'
 import ListItem from '~/components/ListItem'
 
 const Theme = ({ open, onClose }: { open: boolean; onClose: () => void }) => {
   const { setAppBarTitle } = useContext(AppBarTitleContext)
   const { setTheme } = useContext(ThemeContext)
+  const { t } = useTranslation()
 
   useEffect(() => {
-    setAppBarTitle('Select theme')
-  }, [])
+    setAppBarTitle(t('appearance.selectTheme'))
+  }, [setAppBarTitle, t])
 
-  const handleClick = (mode: PaletteMode) => {
+  const handleClick = (mode: ThemePreference) => {
     setTheme(mode)
     onClose()
   }
@@ -24,21 +26,27 @@ const Theme = ({ open, onClose }: { open: boolean; onClose: () => void }) => {
   return (
     <Modal open={open} onClose={onClose}>
       <List>
+        <ListItem disablePadding onClick={() => handleClick('system')}>
+          <ListItemButton>
+            <ListItemText primary={t('appearance.themes.system')} />
+          </ListItemButton>
+        </ListItem>
+
         <ListItem disablePadding onClick={() => handleClick('light')}>
           <ListItemButton>
-            <ListItemText primary="Light" />
+            <ListItemText primary={t('appearance.themes.light')} />
           </ListItemButton>
         </ListItem>
 
         <ListItem disablePadding onClick={() => handleClick('dark')}>
           <ListItemButton>
-            <ListItemText primary="Dark" />
+            <ListItemText primary={t('appearance.themes.dark')} />
           </ListItemButton>
         </ListItem>
 
         <ListItem disablePadding onClick={() => handleClick('black')}>
           <ListItemButton>
-            <ListItemText primary="Black" />
+            <ListItemText primary={t('appearance.themes.black')} />
           </ListItemButton>
         </ListItem>
       </List>