+++ src/hooks/index.ts
@@ -1,5 +1,4 @@
 // export * from "./useEventListener";
 export * from './useHover'
-export * from './useInterval'
 export * from './useLocalStorage'
 export * from './useMediaQuery'